Feature/reservation (#37)
* group seats in rows

* remove unused

* implement seat reservation

* posting filmShows

* add tests

+++ src/main/java/com/kinoticket/backend/repositories/FilmShowSeatRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @Repository
 public interface FilmShowSeatRepository extends CrudRepository<FilmShowSeat, Long> {
 
-    List<FilmShowSeat> findByFilmShow_id(int filmshow_id);
+    List<FilmShowSeat> findByFilmShow_id(long filmshow_id);
 
     Optional<FilmShowSeat> findBySeat_idAndFilmShow_id(long seat_id, long filmshow_id);
 }
